Insert temporary site link
Adds the correct site link to the app, but note that Neocities does not allow IFrames so it is not functional, thus this URL will be changed again at a later date. I would use GitHub's URLs, but the only way for it to not break YTm15 is to make it Entinty1999.github,io, or use an alt and risk being flagged again.